:refactor: index and home page
